# src/sample1/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: SampleAPI
  description: goa3 sample code.
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /api/v1/admin/users:
    get:
      tags:
      - Admin
      summary: admin list user Admin
      description: List all stored users
      operationId: Admin#admin list user
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminGoa2SampleAdminUserResponseCollection'
      schemes:
      - http
    post:
      tags:
      - Admin
      summary: admin create user Admin
      description: Add new user and return its ID.
      operationId: Admin#admin create user
      parameters:
      - name: Admin Create UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminAdminCreateUserRequestBody'
          required:
          - id
          - name
          - email
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /api/v1/admin/users/{id}:
    get:
      tags:
      - Admin
      summary: admin get user Admin
      description: Show user by ID
      operationId: Admin#admin get user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminAdminGetUserResponseBody'
      schemes:
      - http
    put:
      tags:
      - Admin
      summary: admin update user Admin
      description: Update user item.
      operationId: Admin#admin update user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: User id
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Admin Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminAdminUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminAdminUpdateUserResponseBody'
      schemes:
      - http
    delete:
      tags:
      - Admin
      summary: admin delete user Admin
      description: Delete user by id.
      operationId: Admin#admin delete user
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "204":
          description: No Content response.
      schemes:
      - http
  /api/v1/users:
    get:
      tags:
      - Users
      summary: list user Users
      description: List all stored users
      operationId: Users#list user
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersGoa2SampleUserResponseCollection'
      schemes:
      - http
    post:
      tags:
      - Users
      summary: create user Users
      description: Add new user and return its ID.
      operationId: Users#create user
      parameters:
      - name: Create UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersCreateUserRequestBody'
          required:
          - id
          - name
          - email
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: get user Users
      description: Show user by ID
      operationId: Users#get user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersGetUserResponseBody'
      schemes:
      - http
    put:
      tags:
      - Users
      summary: update user Users
      description: Update user item.
      operationId: Users#update user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: User id
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersUpdateUserResponseBody'
      schemes:
      - http
    delete:
      tags:
      - Users
      summary: delete user Users
      description: Delete user by id.
      operationId: Users#delete user
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "204":
          description: No Content response.
      schemes:
      - http
...
definitions:
  AdminAdminCreateUserRequestBody:
    title: AdminAdminCreateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Consequuntur ipsam ut in adipisci est.
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: Ipsam reiciendis qui.
    example:
      email: Quibusdam illo repudiandae ad assumenda voluptatum neque.
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: Itaque nobis dolores accusamus qui non.
    required:
    - id
    - name
    - email
  AdminAdminGetUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.admin.user+json; view=default'
    type: object
    properties:
      created_at:
        type: string
        example: Non modi esse voluptates.
      deleted_at:
        type: string
        example: Et natus sit repudiandae sunt ea.
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
      updated_at:
        type: string
        example: Est unde aliquid iure maxime nam et.
    description: Admin Get UserResponseBody result type (default view)
    example:
      created_at: Qui rerum sed quaerat sapiente in reiciendis.
      deleted_at: Dolor ratione et veniam non.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Aut neque.
    required:
    - id
    - name
    - email
    - created_at
    - updated_at
  AdminAdminUpdateUserRequestBody:
    title: AdminAdminUpdateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Commodi non dolorem tenetur.
      name:
        type: string
        description: Name of user
        example: Numquam provident illum reiciendis.
    example:
      email: Dolor sint.
      name: Aut numquam omnis.
  AdminAdminUpdateUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.admin.user+json; view=default'
    type: object
    properties:
      created_at:
        type: string
        example: Suscipit doloremque accusantium maxime veniam impedit.
      deleted_at:
        type: string
        example: Magnam repellendus sapiente vero et omnis.
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
      updated_at:
        type: string
        example: Eligendi dignissimos et ex repellendus sed.
    description: Admin Update UserResponseBody result type (default view)
    example:
      created_at: Sed magni sed.
      deleted_at: Nostrum tenetur nesciunt.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Et autem ducimus.
    required:
    - id
    - name
    - email
    - created_at
    - updated_at
  AdminGoa2SampleAdminUserResponseCollection:
    title: 'Mediatype identifier: application/vnd.goa2-sample.admin.user+json; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/Goa2SampleAdminUserResponse'
    description: Admin List UserResponseBody is the result type for an array of Goa2-SampleAdminUserResponse
      (default view)
    example:
    - created_at: Voluptatem tempora beatae animi.
      deleted_at: Itaque eos accusamus soluta.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Cumque sunt.
    - created_at: Voluptatem tempora beatae animi.
      deleted_at: Itaque eos accusamus soluta.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Cumque sunt.
    - created_at: Voluptatem tempora beatae animi.
      deleted_at: Itaque eos accusamus soluta.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Cumque sunt.
    - created_at: Voluptatem tempora beatae animi.
      deleted_at: Itaque eos accusamus soluta.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Cumque sunt.
...
  Goa2SampleAdminUserResponse:
    title: 'Mediatype identifier: application/vnd.goa2-sample.admin.user+json; view=default'
    type: object
    properties:
      created_at:
        type: string
        example: Provident dolor officia in hic nesciunt.
      deleted_at:
        type: string
        example: Dolores earum sequi ut modi.
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
      updated_at:
        type: string
        example: Omnis voluptatibus ut neque aliquid.
    description: User Response (default view)
    example:
      created_at: Quia et dolores ducimus illum.
      deleted_at: Eum quidem sed vero.
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
      updated_at: Aspernatur consequatur culpa cumque voluptas accusantium ipsa.
    required:
    - id
    - name
    - email
    - created_at
    - updated_at
  Goa2SampleUserResponse:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: User Response (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
...
  UsersCreateUserRequestBody:
    title: UsersCreateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Delectus velit exercitationem facere.
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: Libero distinctio.
    example:
      email: Quisquam ullam voluptatem.
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: Et unde natus.
    required:
    - id
    - name
    - email
  UsersGetUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: Get UserResponseBody result type (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
  UsersGoa2SampleUserResponseCollection:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/Goa2SampleUserResponse'
    description: List UserResponseBody is the result type for an array of Goa2-SampleUserResponse
      (default view)
    example:
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
  UsersUpdateUserRequestBody:
    title: UsersUpdateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Inventore vero sed.
      name:
        type: string
        description: Name of user
        example: Temporibus quis totam doloremque aut.
    example:
      email: Ipsam fugiat et iste nihil maiores.
      name: Ducimus aut aut ut neque.
  UsersUpdateUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: Update UserResponseBody result type (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
...